requirements.txt
django==4.2.8
djangorestframework==3.14.0
pillow==10.1.0
djoser==2.2.2
djangorestframework-simplejwt==5.3.1
flake8==6.1.0
isort==5.13.2
factory-boy==3.3.0